Allow configuration stop property.
Indicates how parsing should be stopped. The property when specified
should be an array (otherwise it is ignored) containing strings or
regexp instances to test against each arg. Strings must match exactly
regexp instances must test() against the arg.

The -- pattern is always respected before any custom patterns.

When a custom pattern is encountered the match is stripped from the arg
before appending to the list of unparsed arguments, this allows
collecting the remaining portion of the argument that triggered the stop
to be included as the first item in the unparsed array.