Support variable groups.
Such that groups of *option* arguments (those that expect a value) may
be collected into named objects.

Configuration is done via a *vars* option object whose values may by
regexp or strings, the key is used for the grouping.

This is particularly useful for situations like:

-Dorg.example.field=value

Which can be configured with:

{vars: {properties: '-D'}}

Results in:

{vars: {properties: {'org.example.field': 'value'}}}